Working queue screen

+++ app/src/main/java/com/sebastianvm/musicplayer/model/NowPlayingInfo.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class NowPlayingInfo(
-    val nowPlayingPositionInQueue: Long = -1L,
+    val nowPlayingPositionInQueue: Int = -1,
     val lastRecordedPosition: Long = -1L,
 )
